+++ CHANGELOG.md
@@ -64,6 +64,7 @@ instead of use of M2O bankOrder of InvoicePayment object. Like that, if we remov
 - INVENTORY LINES : some products don't display
 - Partner : Do not clean contacts on partner type change but clean on save
 - Partner : Resolve error on save when modify fields and contactSet at same time. 
+- Interco: Compute price discounted for sale and purchase orders.
 
 ## [5.1.0] - 2018-12-13
 ## Features